Fixed warning. JIRA: CXX-10702
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@86746 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/sra/readers/bam/bamindex.cpp
@@ -1853,7 +1853,7 @@ void CBamAuxIterator::x_InitData()
             ptr = static_cast<const char*>(memchr(ptr, 0, end-ptr));
             if ( ptr ) {
                 // found zero termination
-                m_AuxData.m_ElementCount = ptr-m_AuxData.m_DataPtr;
+                m_AuxData.m_ElementCount = uint32_t(ptr-m_AuxData.m_DataPtr);
                 m_AuxPtr = ptr + 1; // skip zero-termination too
                 return;
             }
